requirements.txt
amqp==5.3.1
APScheduler==3.11.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
babel==2.17.0
billiard==4.2.4
bleach==6.3.0
cbor2==5.8.0
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.5
daphne==4.2.1
dj-user-login-history==1.0.6
Django==6.0.3
django-auditlog==3.4.0
django-colorfield==0.14.0
django-countries==8.2.0
django-environ==0.12.0
django-filter==25.2
django-money==3.5.4
django-multiselectfield==1.0.1
django-summernote==0.8.20.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
exceptiongroup==1.3.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
kombu==5.6.1
msgpack==1.1.2
numpy==2.3.5
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.1.1
prompt_toolkit==3.0.52
py-moneyed==3.0
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.1.0
reportlab==4.4.6
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools>=80.10.2
six==1.17.0
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
wheel>=0.46.2
whitenoise==6.11.0
zope.interface==8.1.1
